+++ eq-author/src/features/design/entities/Section/Design/SectionEditor/SectionEditor.test.js
