+++ eq-author/src/features/design/entities/Section/Design/SectionEditor/index.js
