+++ eq-author/src/features/design/entities/Section/Preview/SectionIntroPreview.js
