+++ eq-author/src/features/design/entities/Section/Preview/SectionIntroPreview.test.js
