+++ eq-author/src/features/design/entities/Section/Preview/index.test.js
